BackEnd/Agent/requirements.txt
langgraph
langchain-google-genai
langchain-core
langchain-community
fastapi
uvicorn
python-dotenv
pydantic
httpx
youtube-transcript-api
sqlmodel
psycopg2-binary
langgraph-checkpoint-postgres
beautifulsoup4
google-generativeai
yt-dlp
python-multipart
typing_extensions
requests
psycopg2-binary
langchain-openai
duckduckgo-search
google-api-python-client